+++ packages/ui/package.json
@@ -55,6 +55,7 @@
     "@radix-ui/react-dialog": "1.0.5",
     "@vanilla-extract/css": "1.14.2",
     "@vanilla-extract/dynamic": "2.1.0",
+    "@vanilla-extract/recipes": "0.5.1",
     "@vanilla-extract/sprinkles": "1.6.1",
     "clsx": "2.1.0",
     "react": "18.2.0",
